added dogbone circle via sketch
added simple sketch circle to selected face with calculated inner corner edges.

Appears to work for any component orientation and any  occurrence or mirror.  Not tested with rootComponent bodies.

Drawback with this approach is the effort finding the approriate profiles.  Because the sketch is based on a face and not a plane, it automatically incorporates existing profiles of the face.  Thus the circle will become not one profile, but 3!!, and the main face profile count will change from 1 to 4.  I think the number of profiles will be 1 + 3n, where n is the number of circles.

One strategy would be to calculate the area of the dogbone segments and filter out all profiles that don't have that area.